Add an example folder with Travis-CI examples
I have placed two example Travis-CI configurations in an examples/
folder. The first is the same as in the README, and the second shows how
to both test and deploy in two stages.

I am hoping that examples might be added for other CI providers as well,
so that users can see different approaches to building and deploying
wheels.